+++ src/reader/linestring.rs
@@ -4,10 +4,9 @@ use crate::common::Dimension;
 use crate::error::{WkbError, WkbResult};
 use crate::reader::coord::Coord;
 use crate::reader::util::{has_srid, ReadBytesExt};
+use crate::reader::HEADER_BYTES;
 use crate::Endianness;
 use geo_traits::LineStringTrait;
-
-const HEADER_BYTES: u64 = 5;
 
 /// A WKB LineString
 ///
@@ -20,11 +19,9 @@ pub struct LineString<'a> {
     /// The number of points in this LineString WKB
     num_points: usize,
 
-    /// This offset will be 0 for a single LineString but it will be non zero for a
-    /// LineString contained within a MultiLineString
-    offset: u64,
+    /// The offset into the buffer where the first coord is located
+    coord_offset: u64,
     dim: Dimension,
-    has_srid: bool,
 }
 
 impl<'a> LineString<'a> {
@@ -34,72 +31,55 @@ impl<'a> LineString<'a> {
     pub(crate) fn try_new(
         buf: &'a [u8],
         byte_order: Endianness,
-        mut offset: u64,
         dim: Dimension,
     ) -> WkbResult<Self> {
-        let has_srid = has_srid(buf, byte_order, offset)?;
-        if has_srid {
-            offset += 4;
-        }
+        let has_srid = has_srid(buf, byte_order)?;
 
+        let num_points_offset = HEADER_BYTES + if has_srid { 4 } else { 0 };
         let mut reader = Cursor::new(buf);
-        reader.set_position(HEADER_BYTES + offset);
+        reader.set_position(num_points_offset);
         let num_points = reader
             .read_u32(byte_order)?
             .try_into()
             .map_err(|e| WkbError::General(format!("Invalid number of points: {}", e)))?;
 
-        let linestring = Self {
+        let coord_offset = num_points_offset + 4; // Skip the 4-byte num_points field
+
+        let mut linestring = Self {
             buf,
             byte_order,
             num_points,
-            offset,
+            coord_offset,
             dim,
-            has_srid,
         };
 
         let expected_end_abs = linestring.coord_offset(num_points as u64);
         if expected_end_abs > buf.len() as u64 {
-            return Self::handle_invalid_buffer_length(
-                linestring.offset,
-                expected_end_abs,
-                buf.len(),
-            );
+            return Self::handle_invalid_buffer_length(expected_end_abs, buf.len());
         }
+
+        linestring.buf = &linestring.buf[0..expected_end_abs as usize];
 
         Ok(linestring)
     }
 
     #[cold]
-    fn handle_invalid_buffer_length(
-        offset: u64,
-        expected_end_abs: u64,
-        buf_len: usize,
-    ) -> WkbResult<Self> {
+    fn handle_invalid_buffer_length(expected_end_abs: u64, buf_len: usize) -> WkbResult<Self> {
         Err(WkbError::General(format!(
-            "Invalid buffer length for LineString: geometry starting at offset {} would end at byte {}, but buffer length is {}.",
-            offset, expected_end_abs, buf_len
+            "Invalid buffer length for LineString: geometry would end at byte {}, but buffer length is {}.",
+            expected_end_abs, buf_len
         )))
     }
 
     /// The number of bytes in this object, including any header
-    ///
-    /// Note that this is not the same as the length of the underlying buffer
+    #[inline]
     pub fn size(&self) -> u64 {
-        // - 1: byteOrder
-        // - 4: wkbType
-        // - 4: numPoints
-        // - 2 * 8 * self.num_points: two f64s for each coordinate
-        let mut header = 1 + 4 + 4;
-        if self.has_srid {
-            header += 4;
-        }
-        header + (self.dim.size() as u64 * 8 * self.num_points as u64)
+        self.buf.len() as u64
     }
 
     /// The offset into this buffer of any given coordinate
     pub fn coord_offset(&self, i: u64) -> u64 {
-        self.offset + 1 + 4 + 4 + (self.dim.size() as u64 * 8 * i)
+        self.coord_offset + (self.dim.size() as u64 * 8 * i)
     }
 
     /// The dimension of this LineString
@@ -113,7 +93,7 @@ impl<'a> LineString<'a> {
     #[inline]
     pub fn coords_slice(&self) -> &'a [u8] {
         let start = self.coord_offset(0) as usize;
-        let end = start + self.dim.size() * 8 * self.num_points;
+        let end = self.coord_offset(self.num_points as u64) as usize;
         &self.buf[start..end]
     }
 
@@ -121,6 +101,12 @@ impl<'a> LineString<'a> {
     #[inline]
     pub fn byte_order(&self) -> Endianness {
         self.byte_order
+    }
+
+    /// Get the underlying buffer of this LineString
+    #[inline]
+    pub fn buf(&self) -> &'a [u8] {
+        self.buf
     }
 }
 
@@ -137,12 +123,9 @@ impl<'a> LineStringTrait for LineString<'a> {
 
     #[inline]
     unsafe fn coord_unchecked(&self, i: usize) -> Self::CoordType<'_> {
-        Coord::new(
-            self.buf,
-            self.byte_order,
-            self.coord_offset(i as u64),
-            self.dim,
-        )
+        let coord_offset = self.coord_offset(i as u64) as usize;
+        let buf = &self.buf[coord_offset..];
+        Coord::new(buf, self.byte_order, self.dim)
     }
 }
 
@@ -159,11 +142,8 @@ impl<'a> LineStringTrait for &LineString<'a> {
 
     #[inline]
     unsafe fn coord_unchecked(&self, i: usize) -> Self::CoordType<'_> {
-        Coord::new(
-            self.buf,
-            self.byte_order,
-            self.coord_offset(i as u64),
-            self.dim,
-        )
+        let coord_offset = self.coord_offset(i as u64) as usize;
+        let buf = &self.buf[coord_offset..];
+        Coord::new(buf, self.byte_order, self.dim)
     }
 }