+++ src/reader/multipolygon.rs
@@ -4,20 +4,17 @@ use crate::common::Dimension;
 use crate::error::{WkbError, WkbResult};
 use crate::reader::polygon::Polygon;
 use crate::reader::util::{has_srid, ReadBytesExt};
+use crate::reader::HEADER_BYTES;
 use crate::Endianness;
 use geo_traits::MultiPolygonTrait;
-
-/// skip endianness and wkb type
-const HEADER_BYTES: u64 = 5;
 
 /// A WKB MultiPolygon
 #[derive(Debug, Clone)]
 pub struct MultiPolygon<'a> {
     /// A Polygon object for each of the internal line strings
     wkb_polygons: Vec<Polygon<'a>>,
-
+    buf: &'a [u8],
     dim: Dimension,
-    has_srid: bool,
 }
 
 impl<'a> MultiPolygon<'a> {
@@ -26,60 +23,46 @@ impl<'a> MultiPolygon<'a> {
         byte_order: Endianness,
         dim: Dimension,
     ) -> WkbResult<Self> {
-        let mut offset = 0;
-        let has_srid = has_srid(buf, byte_order, offset)?;
-        if has_srid {
-            offset += 4;
-        }
+        let has_srid = has_srid(buf, byte_order)?;
+        let num_polygons_offset = HEADER_BYTES + if has_srid { 4 } else { 0 };
 
         let mut reader = Cursor::new(buf);
-        reader.set_position(HEADER_BYTES + offset);
+        reader.set_position(num_polygons_offset);
         let num_polygons = reader
             .read_u32(byte_order)?
             .try_into()
             .map_err(|e| WkbError::General(format!("Invalid number of polygons: {}", e)))?;
 
-        // - 1: byteOrder
-        // - 4: wkbType
-        // - 4: numLineStrings
-        let mut polygon_offset = 1 + 4 + 4;
-        if has_srid {
-            polygon_offset += 4;
-        }
+        let mut polygon_offset = num_polygons_offset + 4;
 
         let mut wkb_polygons = Vec::with_capacity(num_polygons);
         for _ in 0..num_polygons {
-            let polygon = Polygon::try_new(buf, byte_order, polygon_offset, dim)?;
+            let polygon = Polygon::try_new(&buf[polygon_offset as usize..], byte_order, dim)?;
             polygon_offset += polygon.size();
             wkb_polygons.push(polygon);
         }
 
         Ok(Self {
             wkb_polygons,
+            buf: &buf[0..polygon_offset as usize],
             dim,
-            has_srid,
         })
     }
 
     /// The number of bytes in this object, including any header
-    ///
-    /// Note that this is not the same as the length of the underlying buffer
+    #[inline]
     pub fn size(&self) -> u64 {
-        // - 1: byteOrder
-        // - 4: wkbType
-        // - 4: numPolygons
-        let mut header = 1 + 4 + 4;
-        if self.has_srid {
-            header += 4;
-        }
-        self.wkb_polygons
-            .iter()
-            .fold(header, |acc, x| acc + x.size())
+        self.buf.len() as u64
     }
 
     /// The dimension of this MultiPolygon
     pub fn dimension(&self) -> Dimension {
         self.dim
+    }
+
+    /// Get the underlying buffer of this MultiPolygon
+    pub fn buf(&self) -> &'a [u8] {
+        self.buf
     }
 }